Move UHI definitions to a separate file
