+++ bindings/pyroot/pythonizations/python/ROOT/_facade.py
@@ -12,53 +12,7 @@ from ._application import PyROOTApplication
 from ._numbadeclare import _NumbaDeclareDecorator
 
 from ._pythonization import pythonization
-
-## ------------------------------------------------
-## ------------------------------------------------
-# TODO(lukas) Move this to a separate file
-class loc:
-    def __init__(self, value):
-        self.value = value
-        self.offset = 0 
-
-    def __add__(self, add_value):
-        self.offset = self.offset + add_value
-        return self
-
-    def __sub__(self, sub_value):
-        self.offset = self.offset - sub_value
-        return self
-
-    def __call__(self, histogram):
-        return histogram.FindBin(self.value) + self.offset
-
-def underflow(histogram):
-    return 0
-
-def overflow(histogram):
-    return histogram.GetNcells() - 1
-
-class rebin:
-    def __init__(self, ngroup, new_name = ""):
-        self.ngroup = ngroup
-        self.new_name = new_name
-
-    def __call__(self, histogram):
-        return histogram.Rebin(self.ngroup, self.new_name)
-
-class sum:
-    def __call__(self, histogram):
-        return histogram.Integral()
-
-## ------------------------------------------------
-## ------------------------------------------------
-
-def add_uhi_helper(module):
-    module.loc = loc
-    module.underflow = underflow
-    module.overflow = overflow
-    module.rebin = rebin
-    module.sum = sum
+from ._uhi import add_uhi_helper
 
 class PyROOTConfiguration(object):
     """Class for configuring PyROOT"""
